+++ redisenv/__main__.py
@@ -11,6 +11,7 @@ def main():
     cli.group(cluster.cluster).add_command(cluster.create)
 
     cli()
-    
+
+
 if __name__ == "__main__":
     main()
