+++ redisenv/commands/cluster.py
@@ -2,8 +2,7 @@ import sys
 
 import click
 
-from ..env import ClusterHandler
-from ..envhelpers import _default_options, genclusterconf, genclusterspec
+from ..env import ClusterHandler, _default_options
 from ..util import free_ports
 from . import defaultenvname
 
@@ -93,9 +92,11 @@ def create(
         sys.exit(3)
 
     ports = free_ports(nodes, cluster=True)
-    cfg = genclusterconf(ports, redisopts)
+    g = ClusterHandler(ctx.obj.get("DESTDIR"))
 
-    sp = genclusterspec(
+    cfg = g.gen_config(ports, redisopts)
+
+    sp = g.gen_spec(
         name,
         nodes,
         version,
@@ -105,7 +106,6 @@ def create(
         replicas,
     )
 
-    g = ClusterHandler(ctx.obj.get("DESTDIR"))
     if force:
         try:
             g.stop(name)
