+++ redisenv/commands/standalone.py
@@ -2,13 +2,12 @@ import sys
 
 import click
 
-from ..env import STANDALONE_TYPE, EnvironmentHandler
-from ..envhelpers import _default_options, genstandalonespec
+from ..env import STANDALONE_TYPE, StandaloneHandler, _default_options
 from . import defaultenvname
 
 
 def standalone():
-    """for creating an environment based on standalone redis"""
+    """create an environment with standalone redis nodes"""
 
 
 @click.command()
@@ -83,7 +82,10 @@ def create(
     redisopts,
 ):
     """create and start a new environment"""
-    sp = genstandalonespec(
+
+    g = StandaloneHandler(ctx.obj.get("DESTDIR"))
+
+    sp = g.gen_spec(
         name,
         nodes,
         version,
@@ -93,7 +95,6 @@ def create(
         ipv6,
         redisopts,
     )
-    g = EnvironmentHandler(ctx.obj.get("DESTDIR"))
     if force:
         try:
             g.stop(name)
