+++ redisenv/env.py
@@ -2,6 +2,7 @@ import json
 import os
 import subprocess
 from typing import Dict, List, Optional
+from .util import free_ports
 
 import jinja2
 import yaml
@@ -11,6 +12,16 @@ SENTINEL_TYPE = "sentinel"
 STANDALONE_TYPE = "standalone"
 REPLICAOF_TYPE = "replicaof"
 CLUSTER_TYPE = "cluster"
+
+
+_default_options = {
+    "_nodes": 1,
+    "_version": "6.2.8",
+    "_image": "redis",
+    "_ipv6": False,
+    "_docker_host_ip": "172.0.0.1",
+    "_cluster_replicas": 1,
+}
 
 
 class EnvironmentHandler:
@@ -151,10 +162,82 @@ class EnvironmentHandler:
             raise
 
 
+class StandaloneHandler(EnvironmentHandler):
+    """Redis Standalone"""
+
+    def gen_spec(
+        self,
+        name: str,
+        nodes: int = _default_options["_nodes"],
+        version: str = _default_options["_version"],
+        image: str = _default_options["_image"],
+        mounts: List = [],
+        conffile: str = "",
+        ipv6: bool = _default_options["_ipv6"],
+        redisopts: List = [],
+    ) -> Dict:
+        """Generate the environment spec, used in generating the
+        docker-compose configuration.
+        """
+        d = {"name": name}
+        d["nodes"] = nodes
+        d["version"] = version
+        d["listening_port"] = 6379
+        d["conffile"] = conffile
+        d["ipv6"] = ipv6
+        d["image"] = image
+        d["redisoptions"] = redisopts
+        d["ports"] = free_ports(nodes)
+
+        d["mounts"] = []
+        for m in mounts:
+            d["mounts"].append({"local": m[0], "remote": m[1]})
+
+        return d
+
+
+class ReplicaHandler(EnvironmentHandler):
+    """Redis replicas"""
+
+    def gen_spec(
+        self,
+        name: str,
+        nodes: int = _default_options["_nodes"],
+        version: str = _default_options["_version"],
+        image: str = _default_options["_image"],
+        mounts: list = [],
+        conffile: str = "",
+        ipv6: bool = _default_options["_ipv6"],
+        redisopts: List = [],
+        replicaof: int = -1,
+        dockerhost: str = None,
+    ) -> Dict:
+        """Generate the environment spec, used in generating the
+        docker-compose configuration.
+        """
+        d = {"name": name}
+        d["nodes"] = nodes
+        d["version"] = version
+        d["listening_port"] = 6379
+        d["conffile"] = conffile
+        d["ipv6"] = ipv6
+        d["image"] = image
+        d["docker_host"] = dockerhost
+        d["redisoptions"] = redisopts
+        d["ports"] = free_ports(nodes)
+        d["replicaof"] = replicaof
+
+        d["mounts"] = []
+        for m in mounts:
+            d["mounts"].append({"local": m[0], "remote": m[1]})
+
+        return d
+
+
 class SentinelHandler(EnvironmentHandler):
     """A wrapper, specificatlly for sentinel"""
 
-    def _genconfigs(self, env_name, config_file_content):
+    def _write_configs(self, env_name, config_file_content):
         """Generate the configuration files for sentinel"""
 
         count = 1
@@ -166,15 +249,117 @@ class SentinelHandler(EnvironmentHandler):
             with open(os.path.join(confdest, "sentinel.conf"), "w+") as fp:
                 fp.write(c)
 
+    def gen_spec(
+        self,
+        name: str,
+        nodes: int = _default_options["_nodes"],
+        version: str = _default_options["_version"],
+        image: str = _default_options["_image"],
+        mounts: List = [],
+        redisconf: str = "",
+        redisopts: List = [],
+        ports: List = [],
+    ) -> Dict:
+        """Generate the sentinel environment spec, and conf file."""
+
+        d = {"name": name}
+        d["nodes"] = nodes
+        d["version"] = version
+        d["listening_port"] = 6379
+        d["sentinel_port"] = 26379
+        d["redisconf"] = redisconf
+        d["image"] = image
+        d["redisoptions"] = redisopts
+
+        d["mounts"] = []
+        for m in mounts:
+            d["mounts"].append({"local": m[0], "remote": m[1]})
+
+        if ports != []:
+            d["ports"] = ports
+        else:
+            d["ports"] = free_ports(nodes)
+
+        return d
+
+    def gen_config(
+        self,
+        ports: List,
+        user: str,
+        password: str,
+        sentinelopts: List = [],
+        docker_ip: str = "",
+    ) -> List:
+        here = os.path.join(os.path.dirname(__file__), "templates")
+        fsl = jinja2.FileSystemLoader(searchpath=here)
+        tenv = jinja2.Environment(loader=fsl, trim_blocks=True)
+        tmpl = tenv.get_template("sentinel.conf.tmpl")
+
+        conffiles = []
+        for p in ports[1:]:
+            context = {
+                "dockerhost": docker_ip,
+                "sentinelport": p,
+                "port": ports[0],
+                "sentinel_user": user,
+                "sentinel_password": password,
+                "sentinelopts": sentinelopts,
+            }
+            conffiles.append(tmpl.render(context))
+        return conffiles
+
     def start(self, env_name, config_file_content, config):
         """Generate the sentinel configs, then start it up"""
-        self._genconfigs(env_name, config_file_content)
+        self._write_configs(env_name, config_file_content)
         self._generate(env_name, config, SENTINEL_TYPE)
         self._start(env_name)
 
 
 class ClusterHandler(EnvironmentHandler):
     """A wrapper, specifically for redis clusters"""
+
+    def gen_config(
+        self,
+        ports: List,
+        redisopts: List = [],
+    ) -> Dict:
+        """Generates the cluster configuration file contents,
+        per cluster node"""
+
+        here = os.path.join(os.path.dirname(__file__), "templates")
+        fsl = jinja2.FileSystemLoader(searchpath=here)
+        tenv = jinja2.Environment(loader=fsl, trim_blocks=True)
+        tmpl = tenv.get_template("cluster.conf.tmpl")
+
+        conffiles = {}
+        for p in ports:
+            context = {"port": p, "redisopts": redisopts}
+            conffiles[str(p)] = tmpl.render(context)
+        return conffiles
+
+    def gen_spec(
+        self,
+        name: str,
+        nodes: int = _default_options["_nodes"],
+        version: str = _default_options["_version"],
+        image: str = _default_options["_image"],
+        mounts: List = [],
+        ports: List = [],
+        replicas: int = _default_options["_cluster_replicas"],
+    ):
+        """Generate the docker-compose variables for the specified
+        cluster configuration."""
+        d = {"name": name}
+        d["nodes"] = nodes
+        d["version"] = version
+        d["image"] = image
+        d["mounts"] = []
+        d["ports"] = ports
+        d["replicas"] = replicas
+        for m in mounts:
+            d["mounts"].append({"local": m[0], "remote": m[1]})
+
+        return d
 
     def _get_clustermap(self, env_name, ports: List = []) -> str:
         nodemapfile = os.path.join(self.envdir, env_name, "configs", "nodemap")
@@ -183,7 +368,7 @@ class ClusterHandler(EnvironmentHandler):
                 fp.write(f"127.0.0.1:{p}\n")
         return nodemapfile
 
-    def _genconfigs(self, env_name: str, config_file_content: Dict):
+    def _write_configs(self, env_name: str, config_file_content: Dict):
         """Generate the redis configuration file for the cluster node"""
         for k, v in config_file_content.items():
             confdest = os.path.join(self.envdir, env_name, "configs", k, "redis.conf")
@@ -217,7 +402,7 @@ class ClusterHandler(EnvironmentHandler):
             env_name, config.get("ports"), config.get("replicas")
         )
         nodemapfile = self._get_clustermap(env_name, config.get("ports"))
-        self._genconfigs(env_name, config_file_content)
+        self._write_configs(env_name, config_file_content)
         config["nodemapfile"] = nodemapfile
         config["startscript"] = startscript
         self._generate(env_name, config, CLUSTER_TYPE)
